+++ components/ComboBoxV2/index.js
@@ -1,2 +1,2 @@
-import ComboBox from './Component';
+import ComboBox from './ComboBox';
 module.exports = ComboBox;
